+++ NEWS.md
@@ -1,3 +1,8 @@
+# kim 0.5.305
+* Added the following function(s): holm_adjusted_p_values
+* Updated the following function(s) (and other functions that use them):
+t_test_pairwise, compare_groups
+
 # kim 0.5.303
 * Updated the following function(s) (and other functions that use them):
 floodlight_multi_by_continuous